requirements.txt
streamlit
openai
vanna
psycopg2-binary
